Add routing to /register and /
To be able to implement a register form, the application uses the
router to navigate to /register (Register component) and to / (the
home screen).

npm run ng g component user/current-register
npm run ng g component home-screen

Both components are still empty, and the main component still shows
the current moods. Next up: Moving the code to show the current moods
to the home screen component.