Add form components of the login form
I started to implement the login form as a "template-driven form",
so that I can later show the difference between template-driven and
reactive forms.

Driven by a first test, I have added the input elements for the
username and password and a submit button. Those do not perform
any actions yet.

Do not forget to import the FromsModule in the main application
module, otherwise you will get strange error messages in the
browser.