Login form sends data to CurrentUserService
When a user submits the login form, it sends the data to the
CurrentUserService that I created with

npm run ng generate service user/current-user

The test, right not, only mocks the service. Note that the test
function is wrapped in a call to async(...) - Otherwise, having
an assertion in the promise of whenStable() would not work.