Login form supports validation
The login form now performs some basic validation (email address
must be non-empty and valid, password must be non-empty). The
component disables the submit button and shows validation messages
when the input is not valid.

I implemented this, again, in a test-driven way, but testing for
the validation messages is a little bit cumbersome.