Add register form as reactive form.
Added the register form in a test-driven way: One test to ensure
that all the inputs are there, using TestBed. And more tests to
ensure that the form works correctly, **without** using the TestBed.
So, those tests should be much quicker to execute.

Note: I was not able to test whether the form itself has the correct
binding to [formGroup].