Add confirm password field and validator
There is a new confirm-password field in the register component.
When password and confirm password are not equal, the form is not
valid - I implemented this with a validator on the form group.

Also, there is a new test that checks whether the submit function
is actually called after clicking the submit button. Note the
fixture.detectChanges() call after setting the input fields - this
is necessary to remove the "disabled" state of the submit button.